Limit the extension to TLS 1.3
We already had the understanding that no one will attempt to deploy the
extension for TLS 1.2 and earlier, since it would be plagued with
middlebox compatibility issues that TLS 1.3 already faces.  However, it
was left as supported because there was no cost to supporting it.  On
the list, David Benjamin has pointed out that:
- clients have no way to ask for "compress only in TLS 1.3" policy,
  meaning that they have to assume that a TLS 1.2 connection could
  attempt to use this extension;
- clients have to support certificate compression in TLS 1.2 as a
  result.

This adds complexity for no reason, and the problem is easily solved by
restricting the extension to 1.3 or higher.